# orbag-api/src/main/resources/api.yaml
openapi: 3.0.1
info:
  title: orbag API
  version: v0-alpha
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/search/execute:
    post:
      tags:
      - search-controller
      operationId: execute
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SerializableTable'
      security:
      - JWT: []
  /api/search/executeLater:
    post:
      tags:
      - search-controller
      operationId: executeLater
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExecuteLaterResponse'
      security:
      - JWT: []
...
  /api/search/template/{configurationItemName}:
    get:
      tags:
      - search-controller
      operationId: getSearchTemplate
      parameters:
      - name: configurationItemName
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SearchRequest'
      security:
      - JWT: []
  /api/search/execute/{searchId}.tsv:
    get:
      tags:
      - search-controller
      operationId: exportTsv
      parameters:
      - name: searchId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
      security:
      - JWT: []
...
components:
  schemas:
    ConfigurationItemReference:
      type: object
      properties:
        identifier:
          type: string
        configurationItemType:
          type: string
        displayLabel:
          type: string
        configurationItemTypeDisplayLabel:
          type: string
...
    SerializableColumn:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum:
          - Primitive
          - Reference
        displayLabel:
          type: string
    SerializableRow:
      type: object
      properties:
        fields:
          type: object
          additionalProperties:
            type: object
        tags:
          uniqueItems: true
          type: array
          items:
            type: string
    SerializableTable:
      type: object
      properties:
        columns:
          type: array
          items:
            $ref: '#/components/schemas/SerializableColumn'
        rows:
          type: array
          items:
            $ref: '#/components/schemas/SerializableRow'
    BooleanField:
      type: object
      properties:
        name:
          type: string
        displayLabel:
          type: string
        category:
          type: string
        value:
          type: boolean
        readOnly:
          type: boolean
        changed:
          type: boolean
    ConfigurationItemReferenceField:
      type: object
      properties:
        name:
          type: string
        displayLabel:
          type: string
        category:
          type: string
        value:
          $ref: '#/components/schemas/ConfigurationItemReference'
        readOnly:
          type: boolean
        changed:
          type: boolean
        configurationItemType:
          type: string
    ConfigurationItemReferenceListField:
      type: object
      properties:
        name:
          type: string
        displayLabel:
          type: string
        category:
          type: string
        value:
          type: array
          items:
            $ref: '#/components/schemas/ConfigurationItemReference'
        readOnly:
          type: boolean
        changed:
          type: boolean
        configurationItemType:
          type: string
    EnumField:
      type: object
      properties:
        name:
          type: string
        displayLabel:
          type: string
        category:
          type: string
        value:
          type: string
        readOnly:
          type: boolean
        changed:
          type: boolean
        allowedValues:
          type: array
          items:
            type: string
    NumericField:
      type: object
      properties:
        name:
          type: string
        displayLabel:
          type: string
        category:
          type: string
        value:
          type: integer
          format: int32
        readOnly:
          type: boolean
        changed:
          type: boolean
    SerializableFieldGroup:
      type: object
      properties:
        booleanFields:
          type: array
          items:
            $ref: '#/components/schemas/BooleanField'
        configurationItemReferenceFields:
          type: array
          items:
            $ref: '#/components/schemas/ConfigurationItemReferenceField'
        enumFields:
          type: array
          items:
            $ref: '#/components/schemas/EnumField'
        numericFields:
          type: array
          items:
            $ref: '#/components/schemas/NumericField'
        stringFields:
          type: array
          items:
            $ref: '#/components/schemas/StringField'
        configurationItemReferenceListFields:
          type: array
          items:
            $ref: '#/components/schemas/ConfigurationItemReferenceListField'
    StringField:
      type: object
      properties:
        name:
          type: string
        displayLabel:
          type: string
        category:
          type: string
        value:
          type: string
        readOnly:
          type: boolean
        changed:
          type: boolean
...
    SearchRequest:
      type: object
      properties:
        configurationItemName:
          type: string
        parameters:
          $ref: '#/components/schemas/SerializableFieldGroup'
        resultType:
          type: string
          enum:
          - ROW_REFERENCE
          - HIGHLIGHTED_FIELDS
          - ALL_FIELDS
    ExecuteLaterResponse:
      type: object
      properties:
        searchId:
          type: string
...
  securitySchemes:
    JWT:
      type: http
      description: Authentication via X-Api-Token header
      scheme: bearer
      bearerFormat: JWT