# orbag-server/src/main/java/orbag/server/search/ExecuteLaterResponse.java
package orbag.server.search;

public class ExecuteLaterResponse {

    String searchId;

    public String getSearchId() {
        return searchId;
    }

    public void setSearchId(String searchId) {
        this.searchId = searchId;
    }
}